+++ app/src/main/java/com/mercadolibre/android/point_mainapp_demo/app/view/bluetooth/contracts/BluetoothSettingsEvents.kt
@@ -16,4 +16,6 @@ internal sealed class BluetoothSettingsEvents {
     class DiscoveryDevicesFound(val devices: BluetoothDeviceModel) : BluetoothSettingsEvents()
     class DiscoveryDevicesUpdate(val devices: BluetoothDeviceModel) : BluetoothSettingsEvents()
     class PairingDevicesStatus(val pair: Pair<BluetoothBondState, BluetoothDeviceModel>) : BluetoothSettingsEvents()
+
+    class ConnectDevicesResult(val connectDevices: BluetoothDeviceModel) : BluetoothSettingsEvents()
 }